// src/interfaces/spotify.ts
export interface SinglePlaylistResponse extends PlaylistObjectFull { }
// ...
interface FollowersObject {
    href: string;
    total: number;
}

/**
 * Image Object
 * [](https://developer.spotify.com/web-api/object-model/)
 */
interface ImageObject {
    height?: number;
    url: string;
    width?: number;
}

/**
 * Paging Object wrapper used for retrieving collections from the Spotify API.
 * [](https://developer.spotify.com/web-api/object-model/#paging-object)
 */
export interface PagingObject<T> {
    href: string;
    items: T[];
    limit: number;
    next: string;
    offset: number;
    previous: string;
    total: number;
}
// ...
interface PlaylistBaseObject {
    collaborative: boolean;
    description: string | null;
    external_urls: ExternalUrlObject;
    href: string;
    id: string;
    images: ImageObject[];
    name: string;
    owner: UserObjectPublic;
    public: boolean;
    snapshot_id: string;
    type: 'playlist';
    uri: string;
}

/**
 * Playlist Object Full
 * [](https://developer.spotify.com/web-api/object-model/)
 */
interface PlaylistObjectFull extends PlaylistBaseObject {
    followers: FollowersObject;
    tracks: PagingObject<PlaylistTrackObject>;
}

/**
 * Playlist Object Simplified
 * [](https://developer.spotify.com/web-api/object-model/)
 */
export interface PlaylistObjectSimplified extends PlaylistBaseObject {
    tracks: {
        href: string;
        total: number;
    };
}

/**
 * The Track Object in Playlists
 * [](https://developer.spotify.com/web-api/object-model/)
 */
interface PlaylistTrackObject {
    added_at: string;
    added_by: UserObjectPublic;
    is_local: boolean;
    track: TrackObjectFull | EpisodeObjectFull;
}